feat: (622) Move UnknownKeyNotificationManager to util
